+++ src/parser/source/text.ts
@@ -1,7 +1,7 @@
 import { TextParser } from '../source';
 import { html } from 'typed-dom';
 
-const separator = /[^0-9a-zA-Z\u0080-\uFFFF]|[\u0300-\u036F]|(?:[0-9a-zA-Z][!?]*h|\?h|[0-9a-gi-zA-Z!?])ttps?:|[0-9a-zA-Z@]?@[0-9a-zA-Z]|[、。]/;
+const separator = /[^0-9a-zA-Z\u0080-\uFFFF]|[\u0300-\u036F]|(?:[0-9a-zA-Z][!?]*h|\?h|[0-9a-gi-zA-Z!?])ttps?:|[0-9a-zA-Z@]?@[0-9a-zA-Z]/;
 
 export const text: TextParser = source => {
   if (source.length === 0) return;
@@ -17,16 +17,6 @@ export const text: TextParser = source => {
               return [[html('br')], source.slice(2)];
             default:
               return [[document.createTextNode(source.slice(1, 2))], source.slice(2)];
-          }
-        case '、':
-        case '。':
-        case '！':
-        case '？':
-          switch (source[1]) {
-            case '\n':
-              return [[document.createTextNode(source.slice(0, 1))], source.slice(2)];
-            default:
-              return [[document.createTextNode(source.slice(0, 1))], source.slice(1)];
           }
         case '\n':
           return [[html('span', { class: 'newline' }, ' ')], source.slice(1)];