+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.46.0
 
+- Always open a media link in new tab.
 - Refactoring with some small breaking changes.
 
 ## 0.45.0